pyproject.toml
[project]
name = "aigent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai==1.12.1",
    "pynvim==0.6.0",
    "python-dotenv==1.1.0",
]
